+++ tests/test_prepare_frames_batch.py
@@ -59,11 +59,10 @@ def test_prepare_frames_segments_rectified_frames_in_batches(tmp_path, monkeypat
     assert all(frame.image_path is not None and frame.image_path.exists() for frame in batch.frames)
     assert all(frame.labels_path is not None and frame.labels_path.exists() for frame in batch.frames)
     assert all(frame.mask_path is not None and frame.mask_path.exists() for frame in batch.frames)
-    # Labels are uint8 in RAM but the cached artifact stays int32.
     assert all(frame.labels.dtype == np.uint8 for frame in batch.frames)
     for frame in batch.frames:
         saved = np.load(frame.labels_path)
-        assert saved.dtype == np.int32
+        assert saved.dtype == np.uint8
         assert np.array_equal(saved, frame.labels)
